requirements.txt
pystray~=0.19.4
bottle~=0.12.23
requests~=2.28.1
Pillow~=9.3.0
sv_ttk==2.4
requests==2.28.1
PyGObject~=3.42.2; sys_platform == "linux"

